requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
camelot-py==1.0.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
colorlog==6.9.0
cryptography==45.0.2
dill==0.4.0
distro==1.9.0
docling==2.33.0
docling-core==2.31.1
docling-ibm-models==3.4.3
docling-parse==4.0.1
easyocr==1.7.2
et_xmlfile==2.0.0
filelock==3.18.0
filetype==1.2.0
Flask==3.1.1
fpdf==1.7.2
frozenlist==1.6.0
fsspec==2025.5.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc1
google-api-python-client==2.169.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.31.4
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonlines==3.1.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
latex2mathml==3.78.0
lazy_loader==0.4
litellm==1.70.2
lxml==5.4.0
markdown-it-py==3.0.0
marko==2.1.3
MarkupSafe==3.0.2
mdurl==0.1.2
mpire==2.10.2
mpmath==1.3.0
multidict==6.4.4
multiprocess==0.70.18
networkx==3.4.2
ninja==1.11.1.4
numpy==2.2.6
openai==1.75.0
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pdf2image==1.17.0
pdfminer.six==20250506
pillow==11.2.1
pluggy==1.6.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
pylatexenc==2.10
PyMuPDF==1.25.5
pyparsing==3.2.3
pypdf==3.17.4
PyPDF2==3.0.1
pypdfium2==4.30.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
rtree==1.4.0
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.15.3
semchunk==2.2.2
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sympy==1.14.0
tabulate==0.9.0
tifffile==2025.5.10
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.52.2
typer==0.15.4
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
Werkzeug==3.1.3
XlsxWriter==3.2.3
yarl==1.20.0
zipp==3.21.0
